Add support to contains for use of parent field containing List<Key>

+++ src/com/google/appengine/datanucleus/scostore/FKSetStore.java
@@ -63,7 +63,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
   /* (non-Javadoc)
    * @see org.datanucleus.store.scostore.CollectionStore#add(org.datanucleus.store.ObjectProvider, java.lang.Object, int)
    */
-  public boolean add(final ObjectProvider ownerOP, Object element, int currentSize) {
+  public boolean add(final ObjectProvider op, Object element, int currentSize) {
     if (element == null) {
       // FK sets allow no nulls (since can't have a FK on a null element!)
       throw new NucleusUserException(LOCALISER.msg("056039"));
@@ -71,13 +71,13 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
 
     if (MetaDataUtils.isOwnedRelation(ownerMemberMetaData)) {
       // Register the parent key for the element when owned
-      Key parentKey = EntityUtils.getKeyForObject(ownerOP.getObject(), ownerOP.getExecutionContext());
-      KeyRegistry.getKeyRegistry(ownerOP.getExecutionContext()).registerParentKeyForOwnedObject(element, parentKey);
+      Key parentKey = EntityUtils.getKeyForObject(op.getObject(), op.getExecutionContext());
+      KeyRegistry.getKeyRegistry(op.getExecutionContext()).registerParentKeyForOwnedObject(element, parentKey);
     }
 
     // Make sure that the element is persisted in the datastore (reachability)
-    final Object newOwner = ownerOP.getObject();
-    ExecutionContext ec = ownerOP.getExecutionContext();
+    final Object newOwner = op.getObject();
+    ExecutionContext ec = op.getExecutionContext();
     boolean inserted = validateElementForWriting(ec, element, new FieldValues() {
       public void fetchFields(ObjectProvider esm) {
         // Find the (element) table storing the FK back to the owner
@@ -101,7 +101,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
               MappingConsumer.MAPPING_TYPE_EXTERNAL_FK);
           if (externalFKMapping != null) {
             // The element has an external FK mapping so set the value it needs to use in the INSERT
-            esm.setAssociatedValue(externalFKMapping, ownerOP.getObject());
+            esm.setAssociatedValue(externalFKMapping, op.getObject());
           }
         }
 
@@ -112,15 +112,15 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
           if (currentOwner == null) {
             // No owner, so correct it
             NucleusLogger.PERSISTENCE.info(LOCALISER.msg("056037",
-                ownerOP.toPrintableID(), ownerMemberMetaData.getFullFieldName(), 
+                op.toPrintableID(), ownerMemberMetaData.getFullFieldName(), 
                 StringUtils.toJVMIDString(esm.getObject())));
             esm.replaceFieldMakeDirty(getFieldNumberInElementForBidirectional(esm), newOwner);
           }
-          else if (currentOwner != newOwner && ownerOP.getReferencedPC() == null) {
+          else if (currentOwner != newOwner && op.getReferencedPC() == null) {
             // Owner of the element is neither this container and not being attached
             // Inconsistent owner, so throw exception
             throw new NucleusUserException(LOCALISER.msg("056038",
-                ownerOP.toPrintableID(), ownerMemberMetaData.getFullFieldName(), 
+                op.toPrintableID(), ownerMemberMetaData.getFullFieldName(), 
                 StringUtils.toJVMIDString(esm.getObject()),
                 StringUtils.toJVMIDString(currentOwner)));
           }
@@ -141,7 +141,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
         Object oldOwner = elementOP.provideField(getFieldNumberInElementForBidirectional(elementOP));
         if (oldOwner != newOwner) {
           if (NucleusLogger.PERSISTENCE.isDebugEnabled()) {
-            NucleusLogger.PERSISTENCE.debug(LOCALISER.msg("055009", ownerOP.toPrintableID(),
+            NucleusLogger.PERSISTENCE.debug(LOCALISER.msg("055009", op.toPrintableID(),
                 ownerMemberMetaData.getFullFieldName(), StringUtils.toJVMIDString(element)));
           }
 
@@ -160,7 +160,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
       }
       else {
         // 1-N unidir so update the FK if not set to be contained in the set
-        if (contains(ownerOP, element)) {
+        if (contains(op, element)) {
           return false;
         }
         else {
@@ -168,7 +168,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
           // Keys (and therefore parents) are immutable so we don't need to ever
           // actually update the parent FK, but we do need to check to make sure
           // someone isn't trying to modify the parent FK
-          EntityUtils.checkParentage(element, ownerOP);
+          EntityUtils.checkParentage(element, op);
           return true;
         }
       }
@@ -179,7 +179,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
   /* (non-Javadoc)
    * @see org.datanucleus.store.scostore.CollectionStore#addAll(org.datanucleus.store.ObjectProvider, java.util.Collection, int)
    */
-  public boolean addAll(ObjectProvider ownerOP, Collection coll, int currentSize) {
+  public boolean addAll(ObjectProvider op, Collection coll, int currentSize) {
     if (coll == null || coll.size() == 0) {
       return false;
     }
@@ -188,7 +188,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
     boolean success = false;
     Iterator iter = coll.iterator();
     while (iter.hasNext()) {
-      if (add(ownerOP, iter.next(), -1)) {
+      if (add(op, iter.next(), -1)) {
         success = true;
       }
     }
@@ -199,11 +199,11 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
   /* (non-Javadoc)
    * @see org.datanucleus.store.scostore.CollectionStore#clear(org.datanucleus.store.ObjectProvider)
    */
-  public void clear(ObjectProvider ownerOP) {
+  public void clear(ObjectProvider op) {
     // Find elements present in the datastore and delete them one-by-one
-    // Removal of child should always delete the child with GAE since cannot null the parent
-    ExecutionContext ec = ownerOP.getExecutionContext();
-    Iterator elementsIter = iterator(ownerOP);
+    // Removal of child should always delete the child with GAE since cannot null the parent in owned relations
+    ExecutionContext ec = op.getExecutionContext();
+    Iterator elementsIter = iterator(op);
     if (elementsIter != null) {
       while (elementsIter.hasNext()) {
         Object element = elementsIter.next();
@@ -213,31 +213,15 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
           elementSM.flush();
         }
         else {
-          // Element not yet marked for deletion so go through the normal process
-          ec.deleteObjectInternal(element);
+          if (MetaDataUtils.isOwnedRelation(ownerMemberMetaData)) {
+            // Element not yet marked for deletion so go through the normal process
+            ec.deleteObjectInternal(element);
+          } else {
+            // TODO Null this out (in parent)
+          }
         }
       }
     }
-  }
-
-  /* (non-Javadoc)
-   * @see org.datanucleus.store.scostore.CollectionStore#contains(org.datanucleus.store.ObjectProvider, java.lang.Object)
-   */
-  public boolean contains(ObjectProvider ownerOP, Object element) {
-    ExecutionContext ec = ownerOP.getExecutionContext();
-    if (!validateElementForReading(ec, element)) {
-      return false;
-    }
-
-    // Since we only support owned relationships right now, we can check containment simply 
-    // by looking to see if the element's Key contains the parent Key.
-    Key childKey = EntityUtils.getKeyForObject(element, ec);
-    // Child key can be null if element has not yet been persisted
-    if (childKey == null || childKey.getParent() == null) {
-      return false;
-    }
-    Key parentKey = EntityUtils.getPrimaryKeyAsKey(ec.getApiAdapter(), ownerOP);
-    return childKey.getParent().equals(parentKey);
   }
 
   /* (non-Javadoc)
@@ -262,17 +246,17 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
   /* (non-Javadoc)
    * @see org.datanucleus.store.scostore.CollectionStore#remove(org.datanucleus.store.ObjectProvider, java.lang.Object, int, boolean)
    */
-  public boolean remove(ObjectProvider ownerOP, Object element, int currentSize, boolean allowCascadeDelete) {
+  public boolean remove(ObjectProvider op, Object element, int currentSize, boolean allowCascadeDelete) {
     if (element == null) {
         return false;
     }
-    if (!validateElementForReading(ownerOP.getExecutionContext(), element)) {
+    if (!validateElementForReading(op.getExecutionContext(), element)) {
         return false;
     }
 
     // Find the state manager for the element
     Object elementToRemove = element;
-    ExecutionContext ec = ownerOP.getExecutionContext();
+    ExecutionContext ec = op.getExecutionContext();
     if (ec.getApiAdapter().isDetached(element)) {// User passed in detached object to collection.remove()! {
       // Find an attached equivalent of this detached object (DON'T attach the object itself)
       elementToRemove = ec.findObject(ec.getApiAdapter().getIdForObject(element), true, false, element.getClass().getName());
@@ -293,7 +277,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
     }
 
     // Owner of the element has been changed
-    if (Relation.isBidirectional(relationType) && oldOwner != ownerOP.getObject() && oldOwner != null) {
+    if (Relation.isBidirectional(relationType) && oldOwner != op.getObject() && oldOwner != null) {
       // TODO Handle this. Can't swap owner of elements in GAE
       return false;
     }
@@ -314,7 +298,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
   /* (non-Javadoc)
    * @see org.datanucleus.store.scostore.CollectionStore#removeAll(org.datanucleus.store.ObjectProvider, java.util.Collection, int)
    */
-  public boolean removeAll(ObjectProvider ownerOP, Collection coll, int currentSize) {
+  public boolean removeAll(ObjectProvider op, Collection coll, int currentSize) {
     if (coll == null || coll.size() == 0) {
       return false;
     }
@@ -324,7 +308,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
     boolean success = true;
     Iterator iter = coll.iterator();
     while (iter.hasNext()) {
-      if (remove(ownerOP, iter.next(), -1, true)) {
+      if (remove(op, iter.next(), -1, true)) {
         success = false;
       }
     }
@@ -335,19 +319,19 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
   /* (non-Javadoc)
    * @see org.datanucleus.store.scostore.CollectionStore#update(org.datanucleus.store.ObjectProvider, java.util.Collection)
    */
-  public void update(ObjectProvider ownerOP, Collection coll) {
+  public void update(ObjectProvider op, Collection coll) {
     if (coll == null || coll.isEmpty()) {
-      clear(ownerOP);
+      clear(op);
       return;
     }
 
     // Find existing elements, and remove any that are no longer present
-    Iterator elemIter = iterator(ownerOP);
+    Iterator elemIter = iterator(op);
     Collection existing = new HashSet();
     while (elemIter.hasNext()) {
       Object elem = elemIter.next();
       if (!coll.contains(elem)) {
-        remove(ownerOP, elem, -1, true);
+        remove(op, elem, -1, true);
       } else {
         existing.add(elem);
       }
@@ -359,7 +343,7 @@ public class FKSetStore extends AbstractFKStore implements SetStore {
       while (iter.hasNext()) {
         Object elem = iter.next();
         if (!existing.contains(elem)) {
-          add(ownerOP, elem, 0);
+          add(op, elem, 0);
         }
       }
     }
